Split requirements into hard and soft (#171)
* Split requirements into Hard and Soft Requirements

* Variables and comments rename

* as soft requirement

* rename to distance_to_closest_obstacle

---------
